utility_bots/service_monitor_bot/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
pytest==7.4.3
pytest-asyncio==0.21.1
mypy==1.5.1
black==23.9.1 